+++ src/miner/miner-util.h
@@ -27,8 +27,8 @@ struct Params;
 };
 
 static const bool DEFAULT_GENERATE = false;
-static const int DEFAULT_GENERATE_THREADS_CPU = 0;
-static const int DEFAULT_GENERATE_THREADS_GPU = 0;
+static const uint8_t DEFAULT_GENERATE_THREADS_CPU = 0;
+static const uint8_t DEFAULT_GENERATE_THREADS_GPU = 0;
 
 static const bool DEFAULT_PRINTPRIORITY = false;
